Rename getSizeClass file to useAppearanceAdditions
