+++ packages/experimental/AppearanceAdditions/src/index.ts
@@ -2,4 +2,4 @@ export { NativeAppearanceAdditions } from './NativeAppearanceAdditions';
 export { appearanceAdditions } from './appearanceAdditions';
 export type { AppearanceAdditions, SizeClass as SizeClassIOS } from './NativeAppearanceAdditions.types';
 
-export { useHorizontalSizeClass, useAccessibilityContrastOption, useUserInterfaceLevel } from './getSizeClass';
+export { useHorizontalSizeClass, useAccessibilityContrastOption, useUserInterfaceLevel } from './useAppearanceAdditions';
